fix: land pre-existing uncommitted fixes that were blocking the Vercel build
These were already-correct local fixes from earlier work that never got
committed, so the last pushed commit (and every deploy since) still had
the broken code. Confirmed by today's failed deploy log:
  - ProviderAggregator.ts: duplicate `pseClient` import (TS2300)
  - UserActionMemory.ts: truncated MAX_ACTI -> MAX_ACTIONS (TS2304)
Also included in this batch (same class of bug, verified correct):
  - backtest/types.ts: invalid object key `beatsPSE-Index` needed quoting
  - startServer.ts: truncated PLACEHOLDER_PATTE -> PLACEHOLDER_PATTERNS
  - DirectNseProvider.ts: PSE_QUOTE_URL declared twice with different
    values (NSE and BSE URLs) -> split into NSE_QUOTE_URL/BSE_QUOTE_URL
  - earningsRoutes.ts / earningsSentimentRoutes.ts: same two routes
    (/api/news/:symbol/sentiment, /api/earnings/calendar) were each
    registered twice across both files; de-duplicated, kept the more
    complete handler of each
  - orchestrator.py: bfloat16 instead of float32 model load to fit an
    8GB VPS without OOM

+++ src/pse_ai_engine/orchestrator.py
@@ -41,7 +41,9 @@ def _load_model():
     if not HF_TOKEN:
         raise RuntimeError("HF_TOKEN environment variable is not set")
     login(token=HF_TOKEN)
-    base = AutoModelForCausalLM.from_pretrained(BASE_MODEL, dtype=torch.float32, device_map="cpu")
+    # bfloat16 halves memory vs float32 (~5GB vs ~10GB) — required to fit an 8GB VPS
+    # without OOM. Verified: float32 was OOM-killed by the kernel on this server.
+    base = AutoModelForCausalLM.from_pretrained(BASE_MODEL, dtype=torch.bfloat16, device_map="cpu")
     _model = PeftModel.from_pretrained(base, ADAPTER_REPO, token=HF_TOKEN)
     _tokenizer = AutoTokenizer.from_pretrained(BASE_MODEL, token=HF_TOKEN)
     if _tokenizer.pad_token is None:
